+++ CHANGELOG.md
@@ -3,4 +3,6 @@
 * disable copy of non multiple traits
 * add recipe logic settings
 * fix create multiblock generator infinite rotation
+* add a node to deserialize a recipe to nbt
+* add an option for geckolib renderer to automatically `Schedule State Animation` without the node graph.